+++ PROJECT-CA23/PROJECT-CA23/Models/Dto/AddressDtos/AddressDto.cs
@@ -1,0 +1,15 @@
+﻿using System.ComponentModel.DataAnnotations;
+using PROJECT_CA23.Models.Dto.UserDtos;
+
+namespace PROJECT_CA23.Models.Dto.AddressDto
+{
+    public class AddressDto
+    {
+        public int UserId { get; set; }
+        public UserDto User { get; set; }
+        public string Country { get; set; }
+        public string City { get; set; }
+        public string AddressText { get; set; }
+        public string PostCode { get; set; }
+    }
+}
