+++ PROJECT-CA23/PROJECT-CA23/Models/Dto/LoginDtos/RegisterUserRequest.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace PROJECT_CA23.Models.Dto.LoginDto
+namespace PROJECT_CA23.Models.Dto.LoginDtos
 {
     public class RegisterUserRequest
     {
